rename ark daemonset to ark restic server
