+++ docs/cli-reference/ark_restic.md
@@ -31,4 +31,5 @@ Work with restic repositories
 ### SEE ALSO
 * [ark](ark.md)	 - Back up and restore Kubernetes cluster resources.
 * [ark restic init-repository](ark_restic_init-repository.md)	 - create an encryption key for a restic repository
+* [ark restic server](ark_restic_server.md)	 - Run the ark restic server
