+++ pkg/cmd/ark/ark.go
@@ -33,7 +33,6 @@ import (
 	"github.com/heptio/ark/pkg/cmd/cli/restic"
 	"github.com/heptio/ark/pkg/cmd/cli/restore"
 	"github.com/heptio/ark/pkg/cmd/cli/schedule"
-	"github.com/heptio/ark/pkg/cmd/daemonset"
 	"github.com/heptio/ark/pkg/cmd/server"
 	runplugin "github.com/heptio/ark/pkg/cmd/server/plugin"
 	"github.com/heptio/ark/pkg/cmd/version"
@@ -69,7 +68,6 @@ operations can also be performed as 'ark backup get' and 'ark schedule create'.`
 		delete.NewCommand(f),
 		cliclient.NewCommand(),
 		completion.NewCommand(),
-		daemonset.NewCommand(f),
 		restic.NewCommand(f),
 	)
