+++ pkg/cmd/cli/restic/restic.go
@@ -31,6 +31,7 @@ func NewCommand(f client.Factory) *cobra.Command {
 
 	c.AddCommand(
 		NewInitRepositoryCommand(f),
+		NewServerCommand(f),
 	)
 
 	return c
